backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
yfinance
pandas
sentence-transformers
scikit-learn
python-multipart
pytest
httpx
pytest-asyncio
